fix: keep manager absence server exports async
