+++ apps/webapp/src/app/[locale]/(app)/team/absences/actions.ts
@@ -1,7 +1,6 @@
 "use server";
 
 import { and, asc, count, eq, ilike, inArray, isNull, or, sql } from "drizzle-orm";
-import { DateTime } from "luxon";
 import { headers } from "next/headers";
 import { db } from "@/db";
 import { user } from "@/db/auth-schema";
@@ -20,12 +19,18 @@ import {
 	calculateBusinessDaysWithHalfDays,
 	dateRangesOverlap,
 } from "@/lib/absences/date-utils";
-import type { AbsenceWithCategory, DayPeriod } from "@/lib/absences/types";
+import type { AbsenceWithCategory } from "@/lib/absences/types";
 import { currentTimestamp } from "@/lib/datetime/drizzle-adapter";
 import type { ServerActionResult } from "@/lib/effect/result";
 import { createLogger } from "@/lib/logger";
 import { addCalendarSyncJob } from "@/lib/queue";
-import { mapAbsenceRangeToCanonicalTimestamps } from "../../absences/actions.canonical";
+import {
+	buildCanonicalAbsenceRecordValues,
+	getAbsenceOverlapConflictMessage,
+	managerAbsenceAdvisoryLockKey,
+	normalizeManagerAbsenceListParams,
+	validateRecordAbsenceDateRange,
+} from "./manager-absence-action-helpers";
 import { calculateManagerAbsenceMetrics } from "./manager-absence-metrics";
 import { canActorManageTarget, canUseManagerAbsencePage } from "./manager-absence-permissions";
 import type {
@@ -37,7 +42,6 @@ import type {
 } from "./manager-absence-types";
 
 const logger = createLogger("ManagerAbsenceActions");
-const PAGE_SIZES = [10, 25, 50] as const;
 const ACCESS_ERROR = "Employee not found or not accessible";
 
 type ActorEmployee = typeof employee.$inferSelect & {
@@ -56,102 +60,6 @@ type ListedEmployee = Pick<
 type TargetEmployee = typeof employee.$inferSelect & {
 	user: { name: string; email: string };
 };
-
-type ApprovedCanonicalAbsenceInput = {
-	organizationId: string;
-	employeeId: string;
-	categoryId: string;
-	startDate: string;
-	startPeriod: DayPeriod;
-	endDate: string;
-	endPeriod: DayPeriod;
-	countsAgainstVacation: boolean;
-	createdBy: string;
-};
-
-export function normalizeManagerAbsenceListParams(
-	params: Partial<ManagerAbsenceListParams>,
-): ManagerAbsenceListParams {
-	const pageSize = (PAGE_SIZES as readonly number[]).includes(params.pageSize ?? 0)
-		? (params.pageSize as number)
-		: 25;
-
-	return {
-		search: (params.search ?? "").trim(),
-		page: Number.isInteger(params.page) && params.page && params.page > 0 ? params.page : 1,
-		pageSize,
-		year:
-			Number.isInteger(params.year) && params.year && params.year > 1900
-				? params.year
-				: DateTime.local().year,
-	};
-}
-
-export function validateRecordAbsenceDateRange(input: {
-	startDate: string;
-	startPeriod: DayPeriod;
-	endDate: string;
-	endPeriod: DayPeriod;
-}): string | null {
-	const start = DateTime.fromISO(input.startDate);
-	const end = DateTime.fromISO(input.endDate);
-
-	if (!start.isValid || !end.isValid) {
-		return "Invalid date format";
-	}
-
-	if (start > end) {
-		return "Start date must be before end date";
-	}
-
-	if (input.startDate === input.endDate && input.startPeriod === "pm" && input.endPeriod === "am") {
-		return "Cannot end in the morning if starting in the afternoon on the same day";
-	}
-
-	return null;
-}
-
-export function managerAbsenceAdvisoryLockKey(employeeId: string): string {
-	return `manager_absence:${employeeId}`;
-}
-
-export function getAbsenceOverlapConflictMessage(status: "pending" | "approved"): string {
-	return status === "pending"
-		? "Absence request overlaps with an existing pending request"
-		: "Absence request overlaps with an existing approved absence";
-}
-
-export function buildCanonicalAbsenceRecordValues(input: ApprovedCanonicalAbsenceInput) {
-	const { startAt, endAt } = mapAbsenceRangeToCanonicalTimestamps({
-		startDate: input.startDate,
-		startPeriod: input.startPeriod,
-		endDate: input.endDate,
-		endPeriod: input.endPeriod,
-	});
-
-	return {
-		timeRecord: {
-			organizationId: input.organizationId,
-			employeeId: input.employeeId,
-			recordKind: "absence" as const,
-			startAt,
-			endAt,
-			durationMinutes: Math.max(0, Math.floor((endAt.getTime() - startAt.getTime()) / 60000)),
-			approvalState: "approved" as const,
-			origin: "manual" as const,
-			createdBy: input.createdBy,
-			updatedBy: input.createdBy,
-		},
-		timeRecordAbsence: {
-			organizationId: input.organizationId,
-			recordKind: "absence" as const,
-			absenceCategoryId: input.categoryId,
-			startPeriod: input.startPeriod,
-			endPeriod: input.endPeriod,
-			countsAgainstVacation: input.countsAgainstVacation,
-		},
-	};
-}
 
 export async function getManagerAbsenceEmployees(
 	params: Partial<ManagerAbsenceListParams>,
